+++ psutil/_common.py
@@ -43,7 +43,7 @@ else:
 
 # can't take it from _common.py as this script is imported by setup.py
 PY3 = sys.version_info[0] == 3
-PSUTIL_DEBUG = bool(os.getenv('PSUTIL_DEBUG', 0))
+PSUTIL_DEBUG = bool(os.getenv('PSUTIL_DEBUG'))
 _DEFAULT = object()
 
 __all__ = [
@@ -366,6 +366,25 @@ class TimeoutExpired(Error):
 # ===================================================================
 
 
+# This should be in _compat.py rather than here, but does not work well
+# with setup.py importing this module via a sys.path trick.
+if PY3:
+    if isinstance(__builtins__, dict):  # cpython
+        exec_ = __builtins__["exec"]
+    else:  # pypy
+        exec_ = getattr(__builtins__, "exec")  # noqa
+
+    exec_("""def raise_from(value, from_value):
+    try:
+        raise value from from_value
+    finally:
+        value = None
+    """)
+else:
+    def raise_from(value, from_value):
+        raise value
+
+
 def usage_percent(used, total, round_=None):
     """Calculate percentage usage of 'used' against 'total'."""
     try:
@@ -391,6 +410,15 @@ def memoize(fun):
     1
     >>> foo.cache_clear()
     >>>
+
+    It supports:
+     - functions
+     - classes (acts as a @singleton)
+     - staticmethods
+     - classmethods
+
+    It does NOT support:
+     - methods
     """
     @functools.wraps(fun)
     def wrapper(*args, **kwargs):
@@ -398,7 +426,10 @@ def memoize(fun):
         try:
             return cache[key]
         except KeyError:
-            ret = cache[key] = fun(*args, **kwargs)
+            try:
+                ret = cache[key] = fun(*args, **kwargs)
+            except Exception as err:
+                raise raise_from(err, None)
             return ret
 
     def cache_clear():
@@ -443,11 +474,17 @@ def memoize_when_activated(fun):
             ret = self._cache[fun]
         except AttributeError:
             # case 2: we never entered oneshot() ctx
-            return fun(self)
+            try:
+                return fun(self)
+            except Exception as err:
+                raise raise_from(err, None)
         except KeyError:
             # case 3: we entered oneshot() ctx but there's no cache
             # for this entry yet
-            ret = fun(self)
+            try:
+                ret = fun(self)
+            except Exception as err:
+                raise raise_from(err, None)
             try:
                 self._cache[fun] = ret
             except AttributeError:
@@ -887,7 +924,7 @@ def debug(msg):
     """If PSUTIL_DEBUG env var is set, print a debug message to stderr."""
     if PSUTIL_DEBUG:
         import inspect
-        fname, lineno, func_name, lines, index = inspect.getframeinfo(
+        fname, lineno, _, lines, index = inspect.getframeinfo(
             inspect.currentframe().f_back)
         if isinstance(msg, Exception):
             if isinstance(msg, (OSError, IOError, EnvironmentError)):