+++ psutil/arch/freebsd/sys_socks.c
@@ -62,7 +62,7 @@ psutil_populate_xfiles(void) {
 
 
 struct xfile *
-psutil_get_file_from_sock(void *sock) {
+psutil_get_file_from_sock(kvaddr_t sock) {
     struct xfile *xf;
     int n;